Fixup after repository merge
