+++ lib/Loop.php
@@ -1,415 +1,367 @@
 <?php
 
-namespace Amp\Loop;
+namespace Amp;
 
-use Amp\Loop\Internal\Watcher;
-use AsyncInterop\Loop\Driver;
-use AsyncInterop\Loop\InvalidWatcherException;
+use Amp\Loop\Driver;
+use Amp\Loop\InvalidWatcherException;
+use Amp\Loop\UnsupportedFeatureException;
 
-abstract class Loop extends Driver {
-    // Don't use 1e3 / 1e6, they result in a float instead of int
-    const MILLISEC_PER_SEC = 1000;
-    const MICROSEC_PER_SEC = 1000000;
-
-    /** @var string */
-    private $nextId = "a";
-
-    /** @var \Amp\Loop\Internal\Watcher[] */
-    private $watchers = [];
-
-    /** @var \Amp\Loop\Internal\Watcher[] */
-    private $enableQueue = [];
-
-    /** @var \Amp\Loop\Internal\Watcher[] */
-    private $deferQueue = [];
-
-    /** @var \Amp\Loop\Internal\Watcher[] */
-    private $nextTickQueue = [];
-
-    /** @var callable|null */
-    private $errorHandler;
-
-    /** @var int */
-    private $running = 0;
-
+/**
+ * Accessor to allow global access to the event loop.
+ *
+ * @see \AsyncInterop\Loop\Driver
+ */
+final class Loop
+{
     /**
-     * {@inheritdoc}
+     * @var Driver
      */
-    public function run() {
-        $previous = $this->running;
-        ++$this->running;
-
-        try {
-            while ($this->running > $previous) {
-                if ($this->isEmpty()) {
-                    return;
-                }
-                $this->tick();
-            }
-        } finally {
-            $this->running = $previous;
-        }
-    }
+    private static $driver = null;
 
     /**
-     * {@inheritdoc}
-     */
-    public function stop() {
-        --$this->running > 0 ?: $this->running = 0;
-    }
-
-    /**
-     * @return bool True if no enabled and referenced watchers remain in the loop.
-     */
-    private function isEmpty() {
-        foreach ($this->watchers as $watcher) {
-            if ($watcher->enabled && $watcher->referenced) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Executes a single tick of the event loop.
-     */
-    private function tick() {
-        $this->deferQueue = \array_merge($this->deferQueue, $this->nextTickQueue);
-        $this->nextTickQueue = [];
-
-        $this->activate($this->enableQueue);
-        $this->enableQueue = [];
-
-        try {
-            foreach ($this->deferQueue as $watcher) {
-                if (!isset($this->deferQueue[$watcher->id])) {
-                    continue; // Watcher disabled by another defer watcher.
-                }
-
-                unset($this->watchers[$watcher->id], $this->deferQueue[$watcher->id]);
-
-                $callback = $watcher->callback;
-                $callback($watcher->id, $watcher->data);
-            }
-
-            $this->dispatch(empty($this->nextTickQueue) && empty($this->enableQueue) && $this->running);
-
-        } catch (\Throwable $exception) {
-            if (null === $this->errorHandler) {
-                throw $exception;
-            }
-
-            $errorHandler = $this->errorHandler;
-            $errorHandler($exception);
-        } catch (\Exception $exception) { // @todo Remove when PHP 5.x support is no longer needed.
-            if (null === $this->errorHandler) {
-                throw $exception;
-            }
-
-            $errorHandler = $this->errorHandler;
-            $errorHandler($exception);
-        }
-    }
-
-    /**
-     * Dispatches any pending read/write, timer, and signal events.
+     * Run the event loop and optionally execute a callback within the scope of it.
      *
-     * @param bool $blocking
-     */
-    abstract protected function dispatch($blocking);
-
-    /**
-     * Activates (enables) all the given watchers.
+     * The loop MUST continue to run until it is either stopped explicitly, no referenced watchers exist anymore, or an
+     * exception is thrown that cannot be handled. Exceptions that cannot be handled are exceptions thrown from an
+     * error handler or exceptions that would be passed to an error handler but none exists to handle them.
      *
-     * @param \Amp\Loop\Internal\Watcher[] $watchers
-     */
-    abstract protected function activate(array $watchers);
-
-    /**
-     * Deactivates (disables) the given watcher.
+     * @param callable|null $callback The callback to execute.
      *
-     * @param \Amp\Loop\Internal\Watcher $watcher
+     * @return void
      */
-    abstract protected function deactivate(Watcher $watcher);
-
-    /**
-     * {@inheritdoc}
-     */
-    public function defer(callable $callback, $data = null) {
-        $watcher = new Watcher;
-        $watcher->type = Watcher::DEFER;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->nextTickQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function delay($delay, callable $callback, $data = null) {
-        $delay = (int) $delay;
-
-        if ($delay < 0) {
-            throw new \InvalidArgumentException("Delay must be greater than or equal to zero");
+    public static function run(callable $callback = null)
+    {
+        if ($callback) {
+            self::$driver->defer($callback);
         }
 
-        $watcher = new Watcher;
-        $watcher->type = Watcher::DELAY;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->value = $delay;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->enableQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
+        self::$driver->run();
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function repeat($interval, callable $callback, $data = null) {
-        $interval = (int) $interval;
-
-        if ($interval < 0) {
-            throw new \InvalidArgumentException("Interval must be greater than or equal to zero");
-        }
-
-        $watcher = new Watcher;
-        $watcher->type = Watcher::REPEAT;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->value = $interval;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->enableQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function onReadable($stream, callable $callback, $data = null) {
-        $watcher = new Watcher;
-        $watcher->type = Watcher::READABLE;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->value = $stream;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->enableQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function onWritable($stream, callable $callback, $data = null) {
-        $watcher = new Watcher;
-        $watcher->type = Watcher::WRITABLE;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->value = $stream;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->enableQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
-    }
-
-    /**
-     * {@inheritdoc}
+     * Retrieve the event loop driver that is in scope.
      *
-     * @throws \AsyncInterop\Loop\UnsupportedFeatureException If the pcntl extension is not available.
-     * @throws \RuntimeException If creating the backend signal handler fails.
+     * @return Driver|null
      */
-    public function onSignal($signo, callable $callback, $data = null) {
-        $watcher = new Watcher;
-        $watcher->type = Watcher::SIGNAL;
-        $watcher->id = $this->nextId++;
-        $watcher->callback = $callback;
-        $watcher->value = $signo;
-        $watcher->data = $data;
-
-        $this->watchers[$watcher->id] = $watcher;
-        $this->enableQueue[$watcher->id] = $watcher;
-
-        return $watcher->id;
+    public static function get()
+    {
+        return self::$driver;
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function enable($watcherIdentifier) {
-        if (!isset($this->watchers[$watcherIdentifier])) {
-            throw new InvalidWatcherException($watcherIdentifier, "Cannot enable an invalid watcher identifier: '{$watcherIdentifier}'");
-        }
-
-        $watcher = $this->watchers[$watcherIdentifier];
-
-        if ($watcher->enabled) {
-            return; // Watcher already enabled.
-        }
-
-        $watcher->enabled = true;
-
-        switch ($watcher->type) {
-            case Watcher::DEFER:
-                $this->nextTickQueue[$watcher->id] = $watcher;
-                break;
-
-            default:
-                $this->enableQueue[$watcher->id] = $watcher;
-                break;
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function disable($watcherIdentifier) {
-        if (!isset($this->watchers[$watcherIdentifier])) {
-            return;
-        }
-
-        $watcher = $this->watchers[$watcherIdentifier];
-
-        if (!$watcher->enabled) {
-            return; // Watcher already disabled.
-        }
-
-        $watcher->enabled = false;
-        $id = $watcher->id;
-
-        switch ($watcher->type) {
-            case Watcher::DEFER:
-                if (isset($this->nextTickQueue[$id])) {
-                    // Watcher was only queued to be enabled.
-                    unset($this->nextTickQueue[$id]);
-                } else {
-                    unset($this->deferQueue[$id]);
-                }
-                break;
-
-            default:
-                if (isset($this->enableQueue[$id])) {
-                    // Watcher was only queued to be enabled.
-                    unset($this->enableQueue[$id]);
-                } else {
-                    $this->deactivate($watcher);
-                }
-                break;
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function cancel($watcherIdentifier) {
-        $this->disable($watcherIdentifier);
-        unset($this->watchers[$watcherIdentifier]);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function reference($watcherIdentifier) {
-        if (!isset($this->watchers[$watcherIdentifier])) {
-            throw new InvalidWatcherException($watcherIdentifier, "Cannot reference an invalid watcher identifier: '{$watcherIdentifier}'");
-        }
-
-        $this->watchers[$watcherIdentifier]->referenced = true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function unreference($watcherIdentifier) {
-        if (!isset($this->watchers[$watcherIdentifier])) {
-            throw new InvalidWatcherException($watcherIdentifier, "Cannot unreference an invalid watcher identifier: '{$watcherIdentifier}'");
-        }
-
-        $this->watchers[$watcherIdentifier]->referenced = false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setErrorHandler(callable $callback = null) {
-        $previous = $this->errorHandler;
-        $this->errorHandler = $callback;
-        return $previous;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getInfo() {
-        $watchers = [
-            "referenced"   => 0,
-            "unreferenced" => 0,
-        ];
-
-        $defer = $delay = $repeat = $onReadable = $onWritable = $onSignal = [
-            "enabled"  => 0,
-            "disabled" => 0,
-        ];
-
-        foreach ($this->watchers as $watcher) {
-            switch ($watcher->type) {
-                case Watcher::READABLE: $array = &$onReadable; break;
-                case Watcher::WRITABLE: $array = &$onWritable; break;
-                case Watcher::SIGNAL:   $array = &$onSignal; break;
-                case Watcher::DEFER:    $array = &$defer; break;
-                case Watcher::DELAY:    $array = &$delay; break;
-                case Watcher::REPEAT:   $array = &$repeat; break;
-
-                default: throw new \DomainException("Unknown watcher type");
-            }
-
-            if ($watcher->enabled) {
-                ++$array["enabled"];
-
-                if ($watcher->referenced) {
-                    ++$watchers["referenced"];
-                } else {
-                    ++$watchers["unreferenced"];
-                }
-            } else {
-                ++$array["disabled"];
-            }
-        }
-
-        return [
-            "watchers"    => $watchers,
-            "defer"       => $defer,
-            "delay"       => $delay,
-            "repeat"      => $repeat,
-            "on_readable" => $onReadable,
-            "on_writable" => $onWritable,
-            "on_signal"   => $onSignal,
-            "running"     => (bool) $this->running,
-        ];
-    }
-
-    /**
-     * Returns the same array of data as getInfo().
+     * Stop the event loop.
      *
-     * @return array
+     * When an event loop is stopped, it continues with its current tick and exits the loop afterwards. Multiple calls
+     * to stop MUST be ignored and MUST NOT raise an exception.
+     *
+     * @return void
      */
-    public function __debugInfo() {
-        return $this->getInfo();
+    public static function stop()
+    {
+        self::$driver->stop();
+    }
+
+    /**
+     * Defer the execution of a callback.
+     *
+     * The deferred callable MUST be executed before any other type of watcher in a tick. Order of enabling MUST be
+     * preserved when executing the callbacks.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param callable(string $watcherId, mixed $data) $callback The callback to defer. The `$watcherId` will be
+     *     invalidated before the callback call.
+     * @param mixed $data Arbitrary data given to the callback function as the `$data` parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     */
+    public static function defer(callable $callback, $data = null)
+    {
+        return self::$driver->defer($callback, $data);
+    }
+
+    /**
+     * Delay the execution of a callback.
+     *
+     * The delay is a minimum and approximate, accuracy is not guaranteed. Order of calls MUST be determined by which
+     * timers expire first, but timers with the same expiration time MAY be executed in any order.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param int $delay The amount of time, in milliseconds, to delay the execution for.
+     * @param callable(string $watcherId, mixed $data) $callback The callback to delay. The `$watcherId` will be
+     *     invalidated before the callback call.
+     * @param mixed $data Arbitrary data given to the callback function as the `$data` parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     */
+    public static function delay(int $delay, callable $callback, $data = null)
+    {
+        return self::$driver->delay($delay, $callback, $data);
+    }
+
+    /**
+     * Repeatedly execute a callback.
+     *
+     * The interval between executions is a minimum and approximate, accuracy is not guaranteed. Order of calls MUST be
+     * determined by which timers expire first, but timers with the same expiration time MAY be executed in any order.
+     * The first execution is scheduled after the first interval period.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param int $interval The time interval, in milliseconds, to wait between executions.
+     * @param callable(string $watcherId, mixed $data) $callback The callback to repeat.
+     * @param mixed $data Arbitrary data given to the callback function as the `$data` parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     */
+    public static function repeat(int $interval, callable $callback, $data = null)
+    {
+        return self::$driver->repeat($interval, $callback, $data);
+    }
+
+    /**
+     * Execute a callback when a stream resource becomes readable or is closed for reading.
+     *
+     * Warning: Closing resources locally, e.g. with `fclose`, might not invoke the callback. Be sure to `cancel` the
+     * watcher when closing the resource locally. Drivers MAY choose to notify the user if there are watchers on invalid
+     * resources, but are not required to, due to the high performance impact. Watchers on closed resources are
+     * therefore undefined behavior.
+     *
+     * Multiple watchers on the same stream MAY be executed in any order.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param resource $stream The stream to monitor.
+     * @param callable(string $watcherId, resource $stream, mixed $data) $callback The callback to execute.
+     * @param mixed $data Arbitrary data given to the callback function as the `$data` parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     */
+    public static function onReadable($stream, callable $callback, $data = null)
+    {
+        return self::$driver->onReadable($stream, $callback, $data);
+    }
+
+    /**
+     * Execute a callback when a stream resource becomes writable or is closed for writing.
+     *
+     * Warning: Closing resources locally, e.g. with `fclose`, might not invoke the callback. Be sure to `cancel` the
+     * watcher when closing the resource locally. Drivers MAY choose to notify the user if there are watchers on invalid
+     * resources, but are not required to, due to the high performance impact. Watchers on closed resources are
+     * therefore undefined behavior.
+     *
+     * Multiple watchers on the same stream MAY be executed in any order.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param resource $stream The stream to monitor.
+     * @param callable(string $watcherId, resource $stream, mixed $data) $callback The callback to execute.
+     * @param mixed $data Arbitrary data given to the callback function as the `$data` parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     */
+    public static function onWritable($stream, callable $callback, $data = null)
+    {
+        return self::$driver->onWritable($stream, $callback, $data);
+    }
+
+    /**
+     * Execute a callback when a signal is received.
+     *
+     * Warning: Installing the same signal on different instances of this interface is deemed undefined behavior.
+     * Implementations MAY try to detect this, if possible, but are not required to. This is due to technical
+     * limitations of the signals being registered globally per process.
+     *
+     * Multiple watchers on the same signal MAY be executed in any order.
+     *
+     * The created watcher MUST immediately be marked as enabled, but only be activated (i.e. callback can be called)
+     * right before the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param int $signo The signal number to monitor.
+     * @param callable(string $watcherId, int $signo, mixed $data) $callback The callback to execute.
+     * @param mixed $data Arbitrary data given to the callback function as the $data parameter.
+     *
+     * @return string An unique identifier that can be used to cancel, enable or disable the watcher.
+     *
+     * @throws UnsupportedFeatureException If signal handling is not supported.
+     */
+    public static function onSignal(int $signo, callable $callback, $data = null)
+    {
+        return self::$driver->onSignal($signo, $callback, $data);
+    }
+
+    /**
+     * Enable a watcher to be active starting in the next tick.
+     *
+     * Watchers MUST immediately be marked as enabled, but only be activated (i.e. callbacks can be called) right before
+     * the next tick. Callbacks of watchers MUST NOT be called in the tick they were enabled.
+     *
+     * @param string $watcherId The watcher identifier.
+     *
+     * @return void
+     *
+     * @throws InvalidWatcherException If the watcher identifier is invalid.
+     */
+    public static function enable(string $watcherId)
+    {
+        self::$driver->enable($watcherId);
+    }
+
+    /**
+     * Disable a watcher immediately.
+     *
+     * A watcher MUST be disabled immediately, e.g. if a defer watcher disables a later defer watcher, the second defer
+     * watcher isn't executed in this tick.
+     *
+     * Disabling a watcher MUST NOT invalidate the watcher. Calling this function MUST NOT fail, even if passed an
+     * invalid watcher.
+     *
+     * @param string $watcherId The watcher identifier.
+     *
+     * @return void
+     */
+    public static function disable(string $watcherId)
+    {
+        self::$driver->disable($watcherId);
+    }
+
+    /**
+     * Cancel a watcher.
+     *
+     * This will detatch the event loop from all resources that are associated to the watcher. After this operation the
+     * watcher is permanently invalid. Calling this function MUST NOT fail, even if passed an invalid watcher.
+     *
+     * @param string $watcherId The watcher identifier.
+     *
+     * @return void
+     */
+    public static function cancel(string $watcherId)
+    {
+        self::$driver->cancel($watcherId);
+    }
+
+    /**
+     * Reference a watcher.
+     *
+     * This will keep the event loop alive whilst the watcher is still being monitored. Watchers have this state by
+     * default.
+     *
+     * @param string $watcherId The watcher identifier.
+     *
+     * @return void
+     *
+     * @throws InvalidWatcherException If the watcher identifier is invalid.
+     */
+    public static function reference(string $watcherId)
+    {
+        self::$driver->reference($watcherId);
+    }
+
+    /**
+     * Unreference a watcher.
+     *
+     * The event loop should exit the run method when only unreferenced watchers are still being monitored. Watchers
+     * are all referenced by default.
+     *
+     * @param string $watcherId The watcher identifier.
+     *
+     * @return void
+     *
+     * @throws InvalidWatcherException If the watcher identifier is invalid.
+     */
+    public static function unreference(string $watcherId)
+    {
+        self::$driver->unreference($watcherId);
+    }
+
+    /**
+     * Stores information in the loop bound registry.
+     *
+     * This can be used to store loop bound information. Stored information is package private. Packages MUST NOT
+     * retrieve the stored state of other packages. Packages MUST use the following prefix for keys: `vendor.package.`
+     *
+     * @param string $key The namespaced storage key.
+     * @param mixed $value The value to be stored.
+     *
+     * @return void
+     */
+    public static function setState(string $key, $value)
+    {
+        self::$driver->setState($key, $value);
+    }
+
+    /**
+     * Gets information stored bound to the loop.
+     *
+     * Stored information is package private. Packages MUST NOT retrieve the stored state of other packages. Packages
+     * MUST use the following prefix for keys: `vendor.package.`
+     *
+     * @param string $key The namespaced storage key.
+     *
+     * @return mixed The previously stored value or `null` if it doesn't exist.
+     */
+    public static function getState(string $key)
+    {
+        return self::$driver->getState($key);
+    }
+
+    /**
+     * Set a callback to be executed when an error occurs.
+     *
+     * The callback receives the error as the first and only parameter. The return value of the callback gets ignored.
+     * If it can't handle the error, it MUST throw the error. Errors thrown by the callback or during its invocation
+     * MUST be thrown into the `run` loop and stop the driver.
+     *
+     * Subsequent calls to this method will overwrite the previous handler.
+     *
+     * @param callable(\Throwable|\Exception $error)|null $callback The callback to execute. `null` will clear the
+     *     current handler.
+     *
+     * @return callable(\Throwable|\Exception $error)|null The previous handler, `null` if there was none.
+     */
+    public static function setErrorHandler(callable $callback = null)
+    {
+        return self::$driver->setErrorHandler($callback);
+    }
+
+    /**
+     * Retrieve an associative array of information about the event loop driver.
+     *
+     * The returned array MUST contain the following data describing the driver's currently registered watchers:
+     *
+     *     [
+     *         "defer"            => ["enabled" => int, "disabled" => int],
+     *         "delay"            => ["enabled" => int, "disabled" => int],
+     *         "repeat"           => ["enabled" => int, "disabled" => int],
+     *         "on_readable"      => ["enabled" => int, "disabled" => int],
+     *         "on_writable"      => ["enabled" => int, "disabled" => int],
+     *         "on_signal"        => ["enabled" => int, "disabled" => int],
+     *         "enabled_watchers" => ["referenced" => int, "unreferenced" => int],
+     *         "running"          => bool
+     *     ];
+     *
+     * Implementations MAY optionally add more information in the array but at minimum the above `key => value` format
+     * MUST always be provided.
+     *
+     * @return array Statistics about the loop in the described format.
+     */
+    public static function getInfo()
+    {
+        $driver = self::$driver ?: self::get();
+        return $driver->getInfo();
+    }
+
+    /**
+     * Disable construction as this is a static class.
+     */
+    private function __construct()
+    {
+        // intentionally left blank
     }
 }
