+++ phpunit.xml.dist
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.1/phpunit.xsd"
+        xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/5.7/phpunit.xsd"
         backupGlobals="false"
         backupStaticAttributes="false"
         bootstrap="vendor/autoload.php"
@@ -13,11 +13,8 @@
         stopOnFailure="false"
 >
     <testsuites>
-        <testsuite name="Amp">
+        <testsuite name="Main">
             <directory>test</directory>
-        </testsuite>
-        <testsuite name="Amp Loop PHPT">
-            <directory suffix=".phpt">test/phpt</directory>
         </testsuite>
     </testsuites>
     <filter>
